+++ R/api_inner.R
@@ -277,3 +277,8 @@
                  any = .dataset_samples_any)
     fun(hosts(x), datasets)
 }
+
+
+# parse data --------------------------------------------------------------
+
+