+++ R/simplify.R
@@ -385,7 +385,7 @@ getTCGAdata = function(project = NULL,
 
 ##' @title Easily Download TCGA Data by Several Options
 ##' @description TCGA is a very useful database and here we provide this function to
-##' download TCGA (include TCGA Pancan) datasets in human-friendly way. User who are not
+##' download TCGA (include TCGA Pancan) datasets in human-friendly way. Users who are not
 ##' familiar with R operation will benefit from this.
 ##' @details All availble information about datasets of TCGA can access vis `availTCGA()` and
 ##' check with `showTCGA()`.